+++ ontimize/components/form/navigation/o-form-navigation.component.ts
@@ -5,7 +5,7 @@ import { Subscription } from 'rxjs';
 import { OFormLayoutManagerComponent } from '../../../layouts';
 import { NavigationService, ONavigationItem } from '../../../services/navigation.service';
 import { OntimizeService } from '../../../services/ontimize.service';
-import { dataServiceFactory } from '../../../services/data-service.provider';
+import { dataServiceFactory } from '../../../services/factories';
 import { Codes, Util } from '../../../utils';
 import { OFormComponent } from '../o-form.component';
 import { OFormNavigationClass } from './o-form.navigation.class';