+++ ontimize/components/form/toolbar/o-form-toolbar.component.ts
@@ -1,6 +1,19 @@
 import { CommonModule } from '@angular/common';
-import { Component, ElementRef, forwardRef, Inject, Injector, NgModule, OnDestroy, OnInit, ViewChild, ViewContainerRef, ViewEncapsulation } from '@angular/core';
+import {
+  Component,
+  ElementRef,
+  forwardRef,
+  Inject,
+  Injector,
+  NgModule,
+  OnDestroy,
+  OnInit,
+  ViewChild,
+  ViewContainerRef,
+  ViewEncapsulation
+} from '@angular/core';
 import { BehaviorSubject, Observable, Subscription } from 'rxjs';
+
 import { InputConverter } from '../../../decorators';
 import { DialogService, NavigationService, OPermissions, SnackBarService } from '../../../services';
 import { OSharedModule } from '../../../shared';
@@ -244,12 +257,11 @@ export class OFormToolbarComponent implements OnInit, OnDestroy {
     this.handleAcceptEditOperation();
   }
 
-
   public cancelOperation(): void {
     if (this.isDetail) {
       this.onCloseDetail();
-    } else if (!this.isDetail && this.insertMode) {
-      this.onCloseDetail();
+    } else if (this.insertMode) {
+      this.onBack();
     } else {
       this.onReload();
       this._form.setInitialMode();