+++ ontimize/config/app-config.ts
@@ -68,9 +68,13 @@ export interface Config {
   };
 
   applicationLocales?: string[];
+  defaultLocale?: string;
 
   // serviceType [ undefined | '' | class ]: The service type used (Ontimize REST standart, Ontimize REST JEE or custom implementation) in the whole application. By default 'undefined', that is, Ontimize REST standard service.
   serviceType?: any;
+
+  // exportServiceType [ undefined | '' | class ]: The service used for exportation in the whole application. It shold implement `IExportService` interface. By default 'undefined' OntimizeExportService.
+  exportServiceType?: any;
 
   // servicesConfiguration: [Object]: Configuration parameters of application services.
   servicesConfiguration?: Object;